python/client/requirements.txt
altgraph==0.17
grpcio==1.28.1
grpcio-tools==1.28.1
keyboard==0.13.5
macholib==1.14
protobuf==3.11.3
PyInstaller==3.6
pyobjc==6.2
pyobjc-core==6.2
pyobjc-framework-Accounts==6.2
pyobjc-framework-AddressBook==6.2
pyobjc-framework-AdSupport==6.2
pyobjc-framework-AppleScriptKit==6.2
pyobjc-framework-AppleScriptObjC==6.2
pyobjc-framework-ApplicationServices==6.2
pyobjc-framework-AuthenticationServices==6.2
pyobjc-framework-AutomaticAssessmentConfiguration==6.2
pyobjc-framework-Automator==6.2
pyobjc-framework-AVFoundation==6.2
pyobjc-framework-AVKit==6.2
pyobjc-framework-BusinessChat==6.2
pyobjc-framework-CalendarStore==6.2
pyobjc-framework-CFNetwork==6.2
pyobjc-framework-CloudKit==6.2
pyobjc-framework-Cocoa==6.2
pyobjc-framework-Collaboration==6.2
pyobjc-framework-ColorSync==6.2
pyobjc-framework-Contacts==6.2
pyobjc-framework-ContactsUI==6.2
pyobjc-framework-CoreAudio==6.2
pyobjc-framework-CoreAudioKit==6.2
pyobjc-framework-CoreBluetooth==6.2
pyobjc-framework-CoreData==6.2
pyobjc-framework-CoreHaptics==6.2
pyobjc-framework-CoreLocation==6.2
pyobjc-framework-CoreMedia==6.2
pyobjc-framework-CoreMediaIO==6.2
pyobjc-framework-CoreML==6.2
pyobjc-framework-CoreMotion==6.2
pyobjc-framework-CoreServices==6.2
pyobjc-framework-CoreSpotlight==6.2
pyobjc-framework-CoreText==6.2
pyobjc-framework-CoreWLAN==6.2
pyobjc-framework-CryptoTokenKit==6.2
pyobjc-framework-DeviceCheck==6.2
pyobjc-framework-DictionaryServices==6.2
pyobjc-framework-DiscRecording==6.2
pyobjc-framework-DiscRecordingUI==6.2
pyobjc-framework-DiskArbitration==6.2
pyobjc-framework-DVDPlayback==6.2
pyobjc-framework-EventKit==6.2
pyobjc-framework-ExceptionHandling==6.2
pyobjc-framework-ExecutionPolicy==6.2
pyobjc-framework-ExternalAccessory==6.2
pyobjc-framework-FileProvider==6.2
pyobjc-framework-FileProviderUI==6.2
pyobjc-framework-FinderSync==6.2
pyobjc-framework-FSEvents==6.2
pyobjc-framework-GameCenter==6.2
pyobjc-framework-GameController==6.2
pyobjc-framework-GameKit==6.2
pyobjc-framework-GameplayKit==6.2
pyobjc-framework-ImageCaptureCore==6.2
pyobjc-framework-IMServicePlugIn==6.2
pyobjc-framework-InputMethodKit==6.2
pyobjc-framework-InstallerPlugins==6.2
pyobjc-framework-InstantMessage==6.2
pyobjc-framework-Intents==6.2
pyobjc-framework-IOSurface==6.2
pyobjc-framework-iTunesLibrary==6.2
pyobjc-framework-LatentSemanticMapping==6.2
pyobjc-framework-LaunchServices==6.2
pyobjc-framework-libdispatch==6.2
pyobjc-framework-LinkPresentation==6.2
pyobjc-framework-LocalAuthentication==6.2
pyobjc-framework-MapKit==6.2
pyobjc-framework-MediaAccessibility==6.2
pyobjc-framework-MediaLibrary==6.2
pyobjc-framework-MediaPlayer==6.2
pyobjc-framework-MediaToolbox==6.2
pyobjc-framework-Metal==6.2
pyobjc-framework-MetalKit==6.2
pyobjc-framework-ModelIO==6.2
pyobjc-framework-MultipeerConnectivity==6.2
pyobjc-framework-NaturalLanguage==6.2
pyobjc-framework-NetFS==6.2
pyobjc-framework-Network==6.2
pyobjc-framework-NetworkExtension==6.2
pyobjc-framework-NotificationCenter==6.2
pyobjc-framework-OpenDirectory==6.2
pyobjc-framework-OSAKit==6.2
pyobjc-framework-OSLog==6.2
pyobjc-framework-PencilKit==6.2
pyobjc-framework-Photos==6.2
pyobjc-framework-PhotosUI==6.2
pyobjc-framework-PreferencePanes==6.2
pyobjc-framework-PubSub==6.2
pyobjc-framework-PushKit==6.2
pyobjc-framework-Quartz==6.2
pyobjc-framework-QuickLookThumbnailing==6.2
pyobjc-framework-SafariServices==6.2
pyobjc-framework-SceneKit==6.2
pyobjc-framework-ScreenSaver==6.2
pyobjc-framework-ScriptingBridge==6.2
pyobjc-framework-SearchKit==6.2
pyobjc-framework-Security==6.2
pyobjc-framework-SecurityFoundation==6.2
pyobjc-framework-SecurityInterface==6.2
pyobjc-framework-ServiceManagement==6.2
pyobjc-framework-Social==6.2
pyobjc-framework-SoundAnalysis==6.2
pyobjc-framework-Speech==6.2
pyobjc-framework-SpriteKit==6.2
pyobjc-framework-StoreKit==6.2
pyobjc-framework-SyncServices==6.2
pyobjc-framework-SystemConfiguration==6.2
pyobjc-framework-SystemExtensions==6.2
pyobjc-framework-UserNotifications==6.2
pyobjc-framework-VideoSubscriberAccount==6.2
pyobjc-framework-VideoToolbox==6.2
pyobjc-framework-Vision==6.2
pyobjc-framework-WebKit==6.2
python-dotenv==0.12.0
six==1.14.0
